requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
bcrypt==4.0.0
billiard==3.6.4.0
blessed==1.19.1
boto3==1.24.82
botocore==1.27.82
cached-property==1.5.2
celery==5.2.7
cement==2.8.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
cryptography==38.0.1
Deprecated==1.2.13
Django==4.1.1
django-cors==0.1
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
idna==3.4
jmespath==0.10.0
jsonschema==3.2.0
kombu==5.2.4
packaging==21.3
paramiko==2.11.0
pathspec==0.9.0
prompt-toolkit==3.0.31
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==5.4.1
redis==4.3.4
requests==2.26.0
s3transfer==0.6.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.3
termcolor==1.1.0
texttable==1.6.4
urllib3==1.26.12
vine==5.0.0
wcwidth==0.1.9
websocket-client==0.59.0
wrapt==1.14.1
